requirements.txt
alembic==1.0.10
aniso8601==7.0.0
arrow==0.14.2
autoenv==1.0.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Classful==0.14.2
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
fleaker==0.5.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==3.0.0rc7
marshmallow-jsonschema==0.6.0
marshmallow-sqlalchemy==0.16.4
openapi-spec-validator==0.2.7
peewee==3.9.6
pendulum==2.0.4
phonenumbers==8.10.13
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
pytzdata==2019.1
PyYAML==5.1.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
